+++ src/NiteCompiler/CodeAnalysis/Symbols/UnsizedArrayTypeSymbol.cs
@@ -5,7 +5,7 @@ public sealed class UnsizedArrayTypeSymbol : BaseArrayTypeSymbol
 	public override TypeKind TypeKind => TypeKind.UnsizedArray;
 	public override SymbolKind Kind => SymbolKind.Array;
 	public override TypeSymbol ElementsType { get; }
-	public override int? Length => null;
+	public override ulong? Length => null;
 
 	public override bool IsUnsized => true;
